+++ src/Inputs.js
@@ -1,9 +1,15 @@
 import { Console } from "@woowacourse/mission-utils";
+import Lotto from "./Lotto.js";
 
 class Inputs {
   async inputPrice() {
-    const price = await Console.readLineAsync("구입금액을 입력해 주세요.\n");
-    return price;
+    try {
+      const price = await Console.readLineAsync("구입금액을 입력해 주세요.\n");
+      return this.validPrice(price);
+    } catch (error) {
+      Console.print(error.message);
+      return this.inputPrice();
+    }
   }
 
   validPrice(price) {
@@ -15,47 +21,31 @@ class Inputs {
   }
 
   async inputWinningNumbers() {
-    const winningNumbers = await Console.readLineAsync(
-      "당첨 번호를 입력해 주세요.\n"
-    );
-    return winningNumbers;
-  }
-
-  validWinningNumbers(numbers) {
-    const winningNumbers = numbers.split(",").map(Number);
-    if (winningNumbers.length !== 6) {
-      throw new Error("[ERROR] 로또 번호는 6개여야 합니다.");
-    }
-
-    const setNumbers = new Set(winningNumbers);
-    if (setNumbers.size !== 6) {
-      throw new Error("[ERROR] 중복된 숫자를 가질 수 없습니다.");
-    }
-
-    if (winningNumbers.some((num) => typeof num !== "number" || isNaN(num))) {
-      throw new Error(
-        "[ERROR] 로또 번호에 숫자가 아닌 문자나 공백이 포함되어 있습니다."
+    try {
+      const winningNumbers = await Console.readLineAsync(
+        "당첨 번호를 입력해 주세요.\n"
       );
+      const winningLotto = new Lotto(winningNumbers.split(",").map(Number));
+      return winningLotto;
+    } catch (error) {
+      Console.print(error.message);
+      return this.inputWinningNumbers();
     }
-
-    if (
-      winningNumbers.some(
-        (num) => !Number.isInteger(num) || num <= 0 || num > 45
-      )
-    ) {
-      throw new Error("[ERROR] 로또 번호는 1부터 45 사이의 정수여야 합니다.");
-    }
-    return winningNumbers;
   }
 
-  async inputBonusNumber() {
-    const bonusNumber = await Console.readLineAsync(
-      "보너스 번호를 입력해 주세요.\n"
-    );
-    return bonusNumber;
+  async inputBonusNumber(winningLotto) {
+    try {
+      const bonusNumber = await Console.readLineAsync(
+        "보너스 번호를 입력해 주세요.\n"
+      );
+      return this.validBonusNumber(bonusNumber, winningLotto.getNumbers());
+    } catch (error) {
+      Console.print(error.message);
+      return this.inputBonusNumber();
+    }
   }
 
-  validBonusNumber(bonusNumber, winningNumbers) {
+  validBonusNumber(bonusNumber, winningLotto) {
     const numericBonusNumber = Number(bonusNumber);
     if (
       isNaN(numericBonusNumber) ||
@@ -67,12 +57,12 @@ class Inputs {
       throw new Error("[ERROR] 보너스 번호는 1부터 45 사이의 정수여야 합니다.");
     }
 
-    if (winningNumbers.includes(numericBonusNumber)) {
+    if (winningLotto.includes(numericBonusNumber)) {
       throw new Error(
         "[ERROR] 보너스 번호는 당첨 번호와 중복되지 않아야 합니다."
       );
     }
-    return numericPrice;
+    return numericBonusNumber;
   }
 }
